+++ gprofiler/profilers/__init__.py
@@ -5,5 +5,6 @@ from gprofiler.profilers.perf import SystemProfiler
 from gprofiler.profilers.php import PHPSpyProfiler
 from gprofiler.profilers.python import PythonProfiler
 from gprofiler.profilers.ruby import RbSpyProfiler
+from gprofiler.profilers.python_ebpf import PythonEbpfProfiler
 
 __all__ = ["JavaProfiler", "SystemProfiler", "PHPSpyProfiler", "PythonProfiler", "RbSpyProfiler", "DotnetProfiler"]
